Added options to prune orphaned files and labels from the dump.
In case when the repository is prepared for conversion by
destroying unnecessary branches, they tend to end up as orphans,
as VSS 8 seems to be reluctant in actually deleting data
files. It may take many runs of Analyze to get rid of them.
Thus, it is useful to be able to prune them during conversion.

Yet, sometimes orphans are moved into non-orphaned space, for
instance when they are produced by recovering from an archive.
This patch runs an additional pass of action processing to collect
data about moves.

Note also that a very large number of orphans causes SVN to
consume about 200K per revision, as it seems to store a complete
list of files in one folder, if even just one of them changes,
(and all orphans are put into the /orphaned folder).